requirements.txt
langchain
openai
faiss-cpu
tavily-python
python-dotenv
